Resolve test resources relative to package root
References to test resource files are resolved into absolute paths relative to
the package root in the unpacked source tree used to build the tests so that
tests requiring resources can run from `stack test` or `stack repl`. The
resources must be listed in `extra-source-files`. The source tree must remain
unpacked while the tests are run.

We discussed the possibility of doing this with the generated Paths module; that
would work, but it would require that the test resources be listed as
`data-files` that would be installed when the library is installed. To keep the
test resources from leaving the source tree, a small test-specific Paths module
is implemented using Template Haskell.