Remove many traversal schemes, simplify substitution
The definitions from Data.Functor.Traversable which were
were actually used elsewhere could mostly be replaced with
simple uses of functions from Data.Functor.Foldable,
except for cataM and the overcomplicated fixTopDownVisitorM.

fixTopDownVisitorM was only used for patternTopDownVisitorM,
which was only used for substitution.
For that case the tricky mechanism of the first downward pass
returning an (m a -> m a) function for "postprocessing" the result
was only actually used to wrap "local" around a Reader computation.
The net effect is just passing a modified environment downwards
to a recursive call, which doesn't need postprocessing at all.

cataM/fixBottomUpVisitorM was only used for patternBottomUpVisitorM,
which was only used with the monad being Either, for error handling.
It wasn't a good choice for those cases because its sequencing of
operations when applied to Either would end up returning errors from
deeper in the term rather than more immediate errors closer to the
root. These cases were rewritten to favor earlier errors.

With these changes Data.Functor.Traversable could be removed,