Parse attributes through HashMap
Attributes are decoded into a HashMap (by object head) before parsing to avoid
multiple passes over the attribute list. Asymptotic complexity of parsing is
reduced from O(n^2) to O(n). We also ensure that all attributes are well-formed
once so that such a check does not need to be implemented in every instance of
ParseAttributes.