Always parse matching logic constructors the same way
`Kore.Parser.ParserImpl.mlConstructorRemainderParser` is modified so that it
does not take a continuation. Instead, it parses all the matching logic
constructors itself. It takes a level parameter which is used to determine if a
particular constructor should be rejected. Dropping the continuation ensures
that matching logic constructors are always parsed by the same code path, which
fixes the parsing of domain values in aliases.