Kore.Unparser: Separate Unparse from Pretty
I would prefer to remove the Unparse class entirely and replace it with specific
functions implementing unparsing for the appropriate types:

- It is not generally the case that the pattern constructors can be unparsed
  correctly just because their children happen to be unparseable. Implementing
  the various Unparse instances is complicated by this assumption.
- The class isn't really necessary because ultimately we always know the
  concrete type whenever we call `unparse`.

Although the Unparse class is not necessary, changing this would be too
invasive, given that the class is not harmful.

The instances are all moved to Kore.Unparser to avoid circular module
dependencies.