Data.Limit: Remove wrong module name in header
The module header for Haddock had the wrong module name. Actually, Haddock does
not look at the module documentation to find the module name, but at the module
declaration itself; according to the documentation:

> The Module field specifies the current module name. Since the module name can
> be inferred automatically from the source file, it doesn’t affect the output
> of any of the backends. But you might want to include it for any other tools
> that might be parsing these comments without the help of GHC.

Such "other tools" don't seem to exist, but we will keep the Module field
up-to-date anyway.