Kore.Builtin.Int.expectBuiltinDomainInt: Clarify error message
When a non-string-literal domain value was encountered, the error reported that
the domain value is not an integer. That is true, but not very helpful: all
integers are encoded as strings, so a non-string domain value can't be an
integer, but there are many other reasons why it could fail to be an integer,
too.