Simplify Kore.Step.Strategy
The key to simplifying the implementation of strategies was to recognize that,
in the parlance of finite-state machines, the strategy is the _input_, not an
element of the state. The implementation unfolds a tree of configurations based
on the input stream of strategy steps. Meta-strategy features such as the step
limit are implemented on the input stream, not in the strategy language itself;
therefore, the language is simultaneously simpler (fewer productions) and more
expressive (because it inherits some expressivity from the implementation
language). The strategy writer also has detailed control over the structure of
the configuration tree because a node will only be output for one element of the
input strategy stream.