Remove Data.Result
The `Result` type functioned as `Maybe` with an extra constructor `Failure`,
which was only used to simulate the `Left` return of the unification
procedure (which runs in the `ExceptT (UnificationOrSubstitutionError ...)`
monad). Unification is much simpler if the unification cases run in `MaybeT m`
where `m` is `Simplifier` for `\and`-simplification and `ExceptT ...` for
unification.

While implementing this refactoring, a bug was discovered in List and Set
unification which could cause invalid substitutions (of non-function patterns)
to be generated.