Refactor Kore.Attribute.Parser
The module was fairly difficult to understand, but now it is significantly
simpler with no loss of functionality. Most attribute types are now defined
under `Kore.Attribute`.

`StepperAttributes` is refactored to destinguish attributes that are declared in
the Kore definition from those inferred by the backend.

The `ImplicitAttributes` type was used when we needed a `ParseAttributes`
instance, but wanted to ignore all the attributes. However, the name suggested
something to do with implicit Kore definitions; this was not the case, so it was
refactored into `Kore.Attribute.Null`.