Use one solver process per execution
Instead of launching an external solver on demand, the simplifier uses a single
process. This has a small negative impact on the unit tests; when the solver was
launched on-demand, many of the unit tests did not launch the solver at all! The
impact is mitigated by sharing a solver process among some test groups, although
this does require solver queries to be serialized.

The solver is only locked when sending a command, instead of locking the solver
for an entire simplifier thread. Simplification tests that do not need the
solver do not lock it at all.

`unsafeRunSMT` is removed; its only legitimate use was for convenience in the
test suite, but the test framework takes care of setting up and tearing down the
solver now.

Many unit tests requiring the simplifier now use Hedgehog instead of QuickCheck.
Hedgehog is a modern take on QuickCheck that will let us run `SMT` computations
inside a `Property`.