PatternVerifier: Return verified pattern
The pattern verifier is refactored to return a `VerifiedKorePattern`, a
`KorePattern` annotated with certain facts collected during verification. For
now, only the pattern sort is collected.

Refactoring the pattern verifier changes the context of certain verification
error messages. Previously, we were not entirely consistent about the context of
sort errors; depending on the pattern head, the context could indicate the
"inside" or the "outside" of the ill-sorted pattern. Now the context of the sort
error always indicates the "outside" of the ill-sorted pattern. Adding more
specific location information to verifier errors is a subject for future work.

See also:
- https://trello.com/c/d7lfgGkH