Simplify Kore sentence kinds
The kind of sentence types is simplified so that it is parameterized on the
pattern type (`PurePattern` or `KorePattern` generally) instead of the pattern
head (`Pattern` or `UnifiedPattern`). This change allows annotating the patterns
within the sentences. The alternative was to add another parameter (for the
annotations) to each sentence type; removing the `domain` and `variable`
parameters was simpler and also allows each sentence type to be a `Functor`.

The `SentenceAlias` type is simplified to admit only aliases where the left-hand
side of the definition is an application pattern. This is the only valid form of
an alias definition admitted by Kore, but the type was previously more
general. The verifier ensures that the left-hand side's arguments are a
non-repeating sequence of variables and that these variables are the only ones
in scope on the right-hand side.