Select GHC options for RTS performance
By default, we have been setting `+RTS -N -RTS`, which causes the GHC runtime
system (RTS) to spawn one thread per physical processor for the parallel garbage
collector. Having a large number of GC threads contending for memory bandwidth
is not efficient because the default allocation size is small (1 megabyte). The
default allocation size is increased to 32 megabytes (`-A32M`) to make better
use of the parallel garbage collector. An upper bound is placed so that there
will be no more than 8 GC threads (`-qn8`) to ensure that the number of threads
is appropriate to the new default allocation size.