Fully specify instances Recursive, Corecursive
Specifying all class functions for these instances for `PurePattern` and
`KorePattern` allows us to take advantage of specialized implementations in the
parent instances.

(I originally wrote this patch when I was on a performance wild-goose chase. In
the end, this is worth less than 0.5% on performance, but I wasn't going to
throw it away.)