Kore.Step.Simplification.Or: Do not duplicate default simplification
The structure of the module resulted in the default simplification step being
duplicated in two functions. If the default step ever changes, or if we ever
have more than one simplification strategy for \or, both functions would need to
change. By changing the signature of the one simplification step, the
duplication is removed.