BaseStep: Remove variable remapper
Axiom variables are remapped at the beginning of the step. Note that we no
longer record the renaming of bound variables. It did not make sense to do
record bound variable renamings in the first place because bound variable names
are not unique; it is not possible to know which variable the renaming applies
to!