Clean up SortedVariable (#451)
* class SortedVariable: Add member fromVariable

Instances of SortedVariable are required to implement fromVariable, i.e. it must
be possible to represent a Variable as any SortedVariable. Note that the reverse
property does not hold.

* class FreshVariable: Remove member freshVariableFromVariable

The class SortedVariable now provides fromVariable and is solely responsible for
it.

* class SortedVariable: Add member toVariable

* Kore.Step.AxiomPatterns: Add refreshRulePattern

* Test.Kore.Step.AxiomPatterns: Add failing test for refreshRulePattern

* Kore.Step.AxiomPatterns: Implement refreshRulePattern

* Kore.Step.AxiomPatterns: Refactor freeVariables

* Kore.Step.AxiomPatterns: Parameterize type RulePattern over variable

* Kore.Step.AxiomPatterns.refreshRulePattern: Return renamings

* BaseStep: Remove variable remapper

Axiom variables are remapped at the beginning of the step. Note that we no
longer record the renaming of bound variables. It did not make sense to do
record bound variable renamings in the first place because bound variable names
are not unique; it is not possible to know which variable the renaming applies
to!