Un-revert SMT lemma translation (#482)
* kore-exec: Declare SMT lemmas

* evaluateBuiltin: Do not throw for non-normalized arguments

* Builtin.KEqual: Reduce unsolved ==K predicates to Matching Logic

* Condition.Evaluator: Also discharge true predicates

* Generalize refutePredicate

* Translate INT.Int and BOOL.Bool sorts specially

* SmtLemma: Translate more declarations

* Revert "kore-exec: Declare SMT lemmas"

This reverts commit ac1cf2982b31e56ff96531148dd6b69b904ef42d.

We rely on Z3 to reject attempts to redefine builtin symbols; version 4.4.1
rejects such declarations, but versions 4.8.x do not reject them. Therefore,
this commit disables sending new declarations to the solver. At the moment, we
do not send any useful declarations anyway. We are working with the frontend to
develop a blacklist of SMT-LIB symbols shared by all backends; this commit
should be reverted after the blacklist is available.