Kore.Exec.prove: Return all failing results
Returning all failing results is very useful for debugging and exploration.

Instead of this implementation, we may want to consider printing a flat list of
patterns. The result is always a disjunction at the top, but because it is a
list of patterns, that disjunction can quickly become very deeply nested.