evaluateWithDefinitionAxioms: Use matching instead of unification
Using matching instead of unification tends to postpone function application
until there are fewer branching points, which simplifies execution. Before, we
would use unification to apply the rule, but we used a heuristic to reject
common symbolic rewriting solutions. However, there are several (extremely)
common cases which cannot be detected this way. A more robust solution is to use
matching instead of unification in the first place.