Stack 2.1 (#737)
* Revert skipping the generation of haddock for dependencies.

* Dockerfile: Install Stack with pinned script

Instead of installing the latest Stack, use the install script from GitHub,
pinned to a specific version. This will automatically invalidate the Docker
cache when we update Stack, and it will improve build reproducibility.

* Dockerfile: Install Stack via script in repository

Repositories that depend on ours can copy the script into their Docker images.

* scripts/stylish.sh: Set PATH from Stack

This uses the PATH which would be set by `stack exec`, which is more robust than
trying to guess the location where `stylish-haskell` in installed.

* Dockerfile: ADD install-stack.sh to its own directory

* Makefile: Use Stack options consistently

* include.mk: Use K_REPO variable

* stack.yaml: Add sha256, size fields for haskell-src-exts

Without these fields, Stack 2.1 produces an obnoxious warning. With these
fields present, older versions of Stack produce a small, unobtrusive warning.

* Dockerfile: Try to cache Stack dependencies better

We want to cache the dependencies of stylish-haskell, and also to do the build
without profiling, which seems to incorrectly cause cache invalidation.

* scripts/build.sh: Build non-snapshot dependencies after cleaning