Reformat the world with Ormolu
Adopting an automatic formatting tool is an extreme step, fraught with
trade-offs, but it mitigates the risk of accidents; specifically, it reduces the
potential damage caused by an accident. This is a soft suggestion: I'm offering
it for consideration because it's easy to implement, not because I have strong
feelings about it.

If we did adopt a formatting tool, Ormolu is the obvious choice:

- It is the only tool that will format 100% of the code without occasionally
  mangling anything.
- The dictated format is relatively diff-friendly.
- The result is just ugly enough that everyone will hate it a little, but not so
  ugly that anyone will hate it a lot.