Factor out Kore.Step.Simplification.SimplificationType
This doesn't really belong in Kore.Step.Simplification.Data, because most of the
other modules which need it don't use anything else in that module. The closest
place to where it belongs is Kore.Step.Simplification.AndTerms, but doing that
would introduce more circular dependencies. For now, it can live in this little
module all by itself.