scripts/check.sh: Only complain about style of changed files
The style check will now only complain if a file changed in the current pull
request does not comply with the style guide. In other words, we do not run
stylish-haskell on files that are not changed from master. This allows us to
incrementally implement changes to the style guide.