Jenkinsfile: Test stages in sequence
The K and KEVM integration test phases have been unstable. I suspect that the
instability is due to running multiple frontend processes in parallel, which is
known to cause issues. Although it will slow down the Pipeline, we will try
running these stages in sequence; if this improves stability we may take more
invasive steps to run the integration tests in parallel.