runSMT: Stop solver gracefully upon error (#1053)
In normal execution, this does not matter because errors kill the parent process,
dispatching the child solver process. In the unit tests, this leaves unused child
solver processes when a test throws an error. If the number of test errors
exceeds the system's child process limit, the remaining tests will fail: they
will be unable to start new solver processes.