Kore.Step.Step: Postpone checkSubstitutionCoverage (#1092)
checkSubstitutionCoverage was called before checking the the initial and
unification conditions with the SMT solver. As a result, the backend would throw
an error for unifications that failed to produce a covering substitution because
they were \bottom. The check is now postponed until the last moment: immediately
before we try to apply the substitution to finalize the right-hand side of the
applied rule.

Fixes: #1084