functionAnd: Restrict to function symbols
Because we do not evaluate functions during unification, we sometimes have to
emit an Equals predicate between two patterns that we will evaluate later. This
suppresses many errors surrounding missing unification cases. We will restrict
this fallback to the case of And between a function symbol and a
non-simplifiable pattern. This allows us to handle cases like `f(x) ∧ 1` without
suppressing errors for cases involving builtins.

Fixes: #1256