Enable new equation application (#1720)
* maybeEvaluatePattern: Do not add child conditions to side conditions

Adding child conditions to the side conditions without simplifying the result
leads to a denormalized side condition. It is not appropriate to call the
condition simplifier here, so we must not use the child conditions at all. The
child conditions will ultimately be added to the side conditions by the
pattern simplifier, leading to the re-evaluation of any equations which might
require the child conditions.

* SMT.AST: INLINE some parser pieces

* Kore.Equation.Application: Detailed result types

* Do not allow free variables only on the right-hand side

* default.nix: 'threaded' and 'release' flags

* DebugApplyEquation

* DebugAttemptEquation

* Prelude.Kore: Cofree

* Kore.Equation.Application: Log entries print source locations