Rename free rule variables at initialization
Instead of renaming the variables of each semantic rule as it is attempted, the
free variables of each rule are renamed at initialization. The former behavior
exhibits poor performance: the work of renaming is substantially duplicated
because every semantic rule is attempted at every step. There is a small amount
of work to rename the free variables of each _applied_ semantic rule, but the
number of applied rules is small, and this work is not really duplicated.