Materialize haskell.nix project
Materialization saves the part of the Nix expression which would be generated
every time the project is built. This part tracks the package description and
dependencies, so it only changes infrequently. Saving the output of this
step (so that we can skip building it) makes the build faster.

The drawback of this approach is that we must periodically run with
checkMaterialization enabled to detect when the materialization needs to be
updated.