Materialize Nix expressions (#2301)
* Remove support for local overrides

Local overrides ruin binary caching.

* Materialize haskell.nix projects

Materialization saves the part of the Nix expression which would be generated
every time the project is built. This part tracks the package description and
dependencies, so it only changes infrequently. Saving the output of this
step (so that we can skip building it) makes the build faster.

The drawback of this approach is that we must periodically run with
checkMaterialization enabled to detect when the materialization needs to be
updated.

* Separate test and release workflows

* Run release workflow only on push to master

* Add nix/rematerialize.sh

* Run Test workflow on pull requests only

* Add a Stack step to Test workflow