Rewrite of graph traversal used by Kore.Exec and Kore.Reachability.Prove
New traversal models required functionality in data types rather than
using exceptions and monad stack layers to add functionality.

The `graphTraversal` function operates with a list of "steps"
(sequences of "instructions") that take a configuration to a result
containing new configurations.

The traversal is parameterised with a specific transition function
`([Step instr],config) -> m (TransitionResult ([Step instr],config))`
which interprets step results. `TransitionResult` can indicate to
continue with one (`Continuing`) or more (`Branch`) new
configurations, or that the configuration is final, stuck, or should
be stopped.

One reason to stop is when the list of steps runs empty, other reasons
could be to stop execuction on the application of certain rules.

The transition function can be constructed from one that operates on
the primitive instructions, together with an interpretation function
that differs depending on the use case (proving or executing).