Unify SMT and NoSMT; kill MonadExecute
This is the first step towards monomorphisation of the monad transformer
stack used in kore (for improved performance). Previously there were two
separate base monads:

	newtype NoSMT a = NoSMT {getNoSMT :: LoggerT IO a}
	newtype SMT a = SMT {getSMT :: ReaderT SolverSetup (LoggerT IO) a}

The choice of the base monad was determined at runtime. Now we have a
single base monad:

	newtype SMT a = SMT (Maybe SolverSetup -> LoggerT IO a)

This makes it possible to use a monomorphic monad where previously a
polymorphic one was required:

	execute ::
	    KoreSolverOptions ->
	    SmtMetadataTools StepperAttributes ->
	    [SentenceAxiom (TermLike VariableName)] ->
	-   (forall exe. MonadExecute exe => exe r) ->
	+   SMT r ->
	    Main r

Note that this change alone is insufficient to improve performance, as
the rest of the application still uses an abstract MonadSMT context.
However, it lifts one of the obstacles towards that goal.