Simplifier as a base monad; drop SimplifierT
Previously, Simplifier was defined as follows:

	newtype SimplifierT smt a =
	  SimplifierT (StateT SimplifierCache (ReaderT (Env (SimplifierT smt)) smt) a)

	type Simplifier = SimplifierT SMT

The smt type variable was instantiated to either SMT or NoSMT, depending
on whether an external SMT solver was available. Now that SMT and NoSMT
have been unified, we can define Simplifier as a base monad rather than
a transformer:

	newtype Simplifier a =
	  Simplifier (StateT SimplifierCache (ReaderT Env SMT) a)

This definition is simpler and allows us to remove parameterization
from Env.

As a knock-on effect, this forces us to remove some of the unnecessary
polymorphism over MonadSMT in the following functions:
	* makeSerializedModule
	* exec
	* search
	* prove
	* boundedModelCheck
	* matchDisjunction
	* checkFunctions
	* mkSimplifierEnv
	* evalSimplifier
	* evalSimplifierProofs
	* evaluateTerm
	* evaluatePredicate