[kore-rpc] Initial working implementation + test harness stub  (#3146)
This initial implementation sends back dummy values for all the routes. We first need to address #3143 and #3144 to provide correct behavior.

This PR also includes a stub test harness, which is a Python script that launches a new instance of `kore-rpc` and tests the output, generated per the given RPC JSON request, against the golden response. It can be run via `nix build .#test-rpc` and should eventually be added to the CI test run.

There is also a document describing the current JSON RPC API:

https://github.com/runtimeverification/haskell-backend/blob/eea0670eac043684ad8d0ad812eb584d0437007a/docs/2022-07-18-JSON-RPC-Server-API.md