Bypass `hookedSymbols` lookup (#3111)
Problem: In `lookupAxiomSimplifier`, if the axiom identifier for the
given term is an `Axiom.Identifier.Application`, we look for the hook
inside the `hookedSymbols` map.

However, in some cases, the term itself might already carry a hook. If
that's the case, then we don't need to perform a lookup in the
`hookedSymbols` map.

See this comment for context:
https://github.com/runtimeverification/haskell-backend/pull/3184#discussion_r938637613

Solution: check if the term carries a hook first, and only do a lookup
in `hookedSymbols` if it doesn't.