Fix exec depth off-by-one, add unit tests
Fixes #3210

The execution depth information was wrong in cases where a state could not be rewritten further:
In the last execution step, rewriting would yield program state `Remaining` but the execution depth counter was be incremented.

This change uses the prior state in such a case, but only if the state did not actually change. In theory, the remaining state could still have been simplified to something different from the prior. However, in practice each rewriting step ends with a simplifier run so the prior state should already be simplified to the extent possible.