Hoogle targets in Makefile
This adds targets for using a local `hoogle` to the `Makefile`.
`hoogle` must be installed (e.g., `cabal install hoogle`) for this to work.

* `make hoogle-data` creates a hoogle database from haddock documentation
  (will build and run `haddock` unless it has been run before, uses `.stack-work-haddock` work directory)
* `make hoogle` prompts the user for a `Query` and runs it:
```
$ make hoogle
Hoogle 5.0.18.3, https://hoogle.haskell.org/
Query:a -> [a]
Prelude repeat :: a -> [a]
Data.List repeat :: a -> [a]
GHC.List repeat :: a -> [a]
GHC.OldList repeat :: a -> [a]
Data.List.Compat singleton :: a -> [a]
Test.QuickCheck shrinkNothing :: a -> [a]
Test.QuickCheck.Arbitrary shrinkNothing :: a -> [a]
Distribution.Compat.Prelude.Internal repeat :: a -> [a]
Hedgehog.Internal.Prelude repeat :: a -> [a]
GhcPrelude repeat :: a -> [a]
-- plus more results not shown, pass --count=20 to see more
```
* `make hoogle-server` starts a local (https) server on `https://localhost:8080` for ongoing queries.
   - The server stays in the foreground and needs to be terminated with `^C`.
   - This is a `https` server, but using a bogus certificate, so a security warning will pop up.
   - The server provides links to documentation in local haddock files which won't work remotely.