Remove Strategy type from the codebase
`Strategy` was the remainder of a richer type that could express sequential or parallel exploration, but became effectively a list (a binary tree that was traversed in left-to-right order in all use sites).

This change replaces `Strategy a` by `[a]`, and `[Strategy a]` by `[Step a]` (using `type Step a = [a]` from Kore.Exec.GraphTraversal`). Also, some types previously imported into `Kore.Exec.GraphTraversal` are moved there to reduce imports.
Tests for the previous `Strategy` algebraic properties are completely removed as they become utterly trivial.