Remove topHack, avoid otherTopSort when generating, check that we can
Previously we had to have a `topGenerator` to come up with a nullary term of sort `otherTopSort`.
This sort is now filtered out (but left in place for the benefit of other tests).

Also ensuring with a separate test tree that we can generate terms of all sizes and known sorts.