Adjustments to test  data generators (#3262)
* More diverse patterns from Test.Kore.Internal.Pattern.internalPatternGen

`Pattern.fromTermLike` would just copy a given term into the `term` part of a `Pattern`.
`Pattern.parsePatternFromTermLike` decomposes the top-level `\and`s and splits the term into predicate and "actual term" parts.

* Remove topHack, avoid otherTopSort when generating, check that we can

Previously we had to have a `topGenerator` to come up with a nullary term of sort `otherTopSort`.
This sort is now filtered out (but left in place for the benefit of other tests).

Also ensuring with a separate test tree that we can generate terms of all sizes and known sorts.

* Format with fourmolu

* hlint

* avoid term connectives inside function arguments

* inline isFunction

* Format with fourmolu

* Adjust hlint invocation to respect version

Co-authored-by: github-actions <github-actions@github.com>