Experiment caching TermLike hashes
Cache the hash of the `TermLikeF` in the `TermLike`, and extract it
in the `Hashable` instance instead of recursing into the `TermLikeF`.

The precomputed hash in produced using the derived `Hashable` instance
of `TermLikeF`, which will call `hash` on the respective sub-structure
(like `And`, `Or`, ...), wrapped into one of the `TermLikeF`
constructors.  For all such sub-structures, the derived `Hashable`
instance will then recurse into the contents (containing a `TermLike`
again for the case of `TermLike`), thereby using the cached hashes
directly instead of performing a deep recursion.

Partially addresses #3113