Fix test 'Apply with top configuration'
Problem: Test 'Apply with top configuration' started to fail as result
of refactoring (removing BuiltinAndAxiomSimplifier). As a workaround the
expected value in that test was changed, but this value is not supposed to
be changed because any logic has not been changed.

Solution: Incorrect test running was the reason of test failing. Fixing
test running according changes made in the refactoring solves the issue.