Df/caching termlike hashes (#3338)
* Experiment caching TermLike hashes

Cache the hash of the `TermLikeF` in the `TermLike`, and extract it
in the `Hashable` instance instead of recursing into the `TermLikeF`.

The precomputed hash in produced using the derived `Hashable` instance
of `TermLikeF`, which will call `hash` on the respective sub-structure
(like `And`, `Or`, ...), wrapped into one of the `TermLikeF`
constructors.  For all such sub-structures, the derived `Hashable`
instance will then recurse into the contents (containing a `TermLike`
again for the case of `TermLike`), thereby using the cached hashes
directly instead of performing a deep recursion.

Partially addresses #3113

* fixup! Experiment caching TermLike hashes

* Vacuous TermLike

Write a specialized

    from :: TermLike Concrete -> TermLike variable

I was hoping this would remove a number of the new `Hashable`
constraints; unfortunately it was not so successful in that regard.  It
did, however remove quite a lot of `InternalVariable` constraints, so I
guess that's a win.

* Lazy hashing

* Fix for fourmolu

`fourmolu` support for CPP is experimental, limited, and
basically buggy. It's apparently okay to make whole top-level
declarations conditional, so let's do it that way.

* Golden test

Co-authored-by: rv-jenkins <admin@runtimeverification.com>