Reorganize the functions parameters in kore-exec Main (#3485)
Problem: `mainDispatch` function uses `KoreExecOptions` data type but
gets it wrapped with `LocalOptions`. This leads to unnecessary
wrapping-unwrapping. `koreSearch`, `koreProve`, `koreSerialize` and
`koreRun` internally also use `KoreExecOptions` but take as a parameter
`LocalOptions`. `koreProve` and `koreSearch` besides `KoreExecOptions`
accepts as a second parameter which is a part of `KoreExecOptions` and
apparently the second parameter is unnecessary.

Solution: make `KoreExecOptions` as the only parameter  of
`mainDispatch` `koreProve`, `koreSearch`, `koreSerialize`, `koreRun`.
Remove the second parameter of functions `koreProve` and `koreSearch`.