Add validation of debug options
Problem: When we debug equational/rewrite rules with options
`--debug-equation`, `--debug-attempt-equation`,
`--debug-apply-equation`, `--debug-rewrite`, `--debug-attempt-rewrite`,
`--debug-apply-rewrite`, if rule labels set with these options does not
exists in definition we don't want these labels to be silently ignored.
`Kore-exec` should throw error pointing the labels which was not found
in the definition.

Solution: After initialization check if there exist labels set by
command line options in the definition. And if any of them are not
defined in the definition throw the error containing comprehensive
description.