Remove check for stuck states from branch case in prover graph traversal
We previously had an invariant that when execution (or a proof) branches, none of the branch states would be `Stuck`.
Since modifying the prover logic to disallow vacuous claims (#3411) this invariant does not hold any more.

Remove the invariant check and return stuck states within the branch states (they will be discovered and handled in the next step). Also, prune `Proven` states from the branch and terminate if all branch states were already `Proven`.