[#3538] Fix converting substitution to `Map`
Problem: Kore-exec fails with "Cannot convert a denormalized
substitution to a map!".

Solution: When appending one condition to another we can get condition
with denormalized substitution which leads to the error. Therefore if we get
condition with denormalized substitution we need to simplify this
condition before converting substitution to `Map`.