[#3170] Detect if equational rules are unsupported (#3559)
* [#3170] Detect if equational rules are unsupported

Problem: Right now we support equational rules where the LHS has a
function(al) symbol, \\equals or \\ceil at the top. The backend should
abort execution if it detects anything else in this spot and provide an
appropriate error message containing the source location of the rule.

Solution: Match LHS to supported terms. And if it don't match, throw the
`UnsupportedLHS` error.

* [#3170] Add integration test

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>