Remove repl output from golden files for repl script tests
With ghc-9.2.7, the haskeline library does not print to stdout when
`outputStrLn` is called, it puts the string into the terminal directly
(cannot be redirected).
The easiest way to work around this is to remove the repl output from the
expectation file - considering that the repl is probably legacy code anyway.