Hotfix avoid looping on map internalisation (#188)
The `Map` internalisation was looping on maps composed of two variables because we were re-entering the smart constructor with the same arguments again.
I also added rule location information for errors during definition internalisation. This might not have to use a type class (we currently ignore symbol and sort location attributes) but it is not expensive to use one.