Un-do function argument encoding with `\in` for smt-lemmas
Function equations encode the function arguments of the LHS by a
special `\in` clause in the side conditions to ensure that the
equation only applies on calls with defined arguments.

However, this gets in the way of sending equations to the SMT solver
as lemmas, so we un-do the special encoding before attempting to
translate the lemma to an SMT s-expression.

Fixes #3616

Added an integration test with the example from the issue (two inputs)