200 - Collect constraints in equation state, abort execution on `ensures(false)` (#227)
Fixes #200 

* `simplifyPattern` has changed to `evaluatePattern` to avoid confusion
with the LLVM-driven `simplifyPattern`.
* Side conditions (`constraints` in a pattern) are stored during
evaluation of function and simplification rules. When
evaluating/simplifying a predicate, it is temporarily removed from the
stored side conditions. (This prepares for adding an SMT solver and
checking constraints in combination later).
* `ensures(P)` clauses in equations add to the stored side conditions,
the returned pattern contains these new side conditions as well as the
(simplified) original ones.
* When an `ensures` clause is found to be (trivially) false, the
evaluation is pruned.
* This makes the rewrite `stuck` if the simplification is called from
`performRewrite`.
* Simplification of rewrite results before they are returned may change
the result because parts may be stuck. In case of branching, the
execution may continue if all branches but one are pruned, other results
may turn into `stuck`.