disable error for duplicate module (#270)
This PR brings the booster in alignment with the old backend by allowing
new modules with a particular name to replace old modules with the same
name. We may want to separately consider changing both to only allow
existing modules if they are identical to the old module, but I believe
that this should be done as a single step to both backends once we
confirm this is the desired behavior, rather than trying to apply the
change piecemeal when the backends differ in behavior to start with.