Explain a subtlety of `next-states` in `branching` results
In a `branching` result, the user might be thinking that all states in
`next-states` have taken a single rewrite step from `state`.
This is not correct.

If the branching condition together with other predicates in `state`
make the state `\bottom` as a whole, the result will contain the
`term` from `state`and the `predicate` adding the branching condition
to the existing predicates from `state`. This state is in fact `stuck`
but is reported to the caller as a branch to provide all available
information.

Closes #3622