Fix a couple memory leaks (#263)
Blocked on #264 

We were never taking any rewrite steps, so the garbage collector was
never getting called. We also forgot to call kllvm_init, which meant
that some of the memory that ought to have been allocated on the garbage
collected heap in the llvm backend was instead being allocated in the
wrong place.

Finally, we forgot to call `free` on the `char*` returned from
`kore_simplify`.

---------

Co-authored-by: Sam Balco <goodlyrottenapple@gmail.com>