Unpack foundry-bug-report.tar.gz integration test (#288)
Unfortunately, this test is not stable enough to stay as is, since every
tweak commits a new binary blob. However, the test has been useful for
catching errors early and should still run on CI. Hence, we unpack and
adapt the structure of the test to run it via the `runDirectoryTest.sh`
script. Whilst unpacking the test, I've also come across a bug in the
rpc-client, wherein the client would crash when running `git diff`. This
was due to `readProcessWithExitCode` returning a `String` instead of
`ByteString`, which resulted in

```
rpc-client: fd:12: hGetContents: invalid argument (invalid byte sequence)
```

on certain byte arrays in the foundry demo response. Since we switched
to the internal diffing in #284, we can just pretty print the diff
instead.