Simplify `ExecuteResult` in stuck/terminal/cut-point cases to enforce substitution (#293)
Part of #275 
Part of #290 

It turns out, the `kore-rpc-booster` and `kore-rpc-dev` returned
different responses for `execute` requests. The reason is: Kore applies
the final substitution to the configuration and Booster does not; thus,
if `kore-rpc-booster` had not fallen back to Kore, the response is
returned without applying the substitution after rewriting.

This PR partially fixes this behavior by calling `postExecSimplify` in
`kore-rpc-booster` before returning the response to `execute`. Prior to
this, we only simplified `Branching` results, but not `TerminalRule`
etc.

The added integration test amounts to:
> take one rewrite step and learn the fact that two variables in the
configuration are in fact equal

Prior to this PR, we could see two variations of the behavior:
* `kore-rpc-dev` (i.e. the vanilla `kore-rpc`) upon learning the fact
that `X #Equals Y`:
  - substitutes `Y` for `X` in the configuration, eliminating `X`.
- returns `X #Equals Y` as `substitution`, while `predicate` is empty.
* `booster-dev` (and, consequentially, `kore-rpc-booster`, since we hit
a terminal rule and do not fall back to Kore):
- *does not* substitute `Y` for `X`, so both `X` and `Y` stay in the
configuration
- returns `X ==Int Y #Equals true` as `predicate`, while `substitution`
is empty.

Now the substitution is applied by `kore-rpc-booster` too. There's still
a discrepancy between how the substitution is returned in the response:
`kore-rpc-dev` puts it into `substitution` as a nested `#And` of
`#Equals`, and `kore-rpc-booster` returns it as part of `predicate`,
specializing `#Equals` to sort equality (e.g. `==Int`).